[FEATURE] Get keys and secrets from 1password (#5)
* [WIP] 1password ssh key management

Starts working on downloading keys from 1password. This is the start of
an attempt to start phasing out sops for secrets management. Nothing
wrong with sops however I am paying for 1password so may as well make
use of all the features

* [WIP] Continue using the old ssh key for now

Still have a bunch of servers that are using my old ssh key. Put this
back into the ssh config for the moment. Will look to remove from older
servers and discontinue the rsa key